refactor(회원가입, 비밀번호 초기화): 선명한 이미지로 변경
