+++ frontend/src/pages/welcome.js
@@ -1,5 +1,5 @@
 import { useRouter } from 'next/router'
-import React, { useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import theme from '../theme/theme';
 import Image from 'next/image';
 import logo from '../image/login_enheng.png'
@@ -8,6 +8,10 @@ import { CssBaseline, Box, ThemeProvider, Grid,Button, Container, Typography } f
 export default function welcome() {
     const router = useRouter();
     const [remainHeight, setRemainHeight] = useState(window.innerHeight - (143+56) + "px");
+
+    useEffect(() => {
+        setRemainHeight(window.innerHeight - (143+56) + "px")
+    }, [window.innerHeight])
 
     return (
         <ThemeProvider theme={theme}>
@@ -21,8 +25,8 @@ export default function welcome() {
             }}
         >
             <div style={{ width: '100%', textAlign: 'center', marginTop: `calc(${remainHeight} * 0.45)` }}>
-            <Typography sx={{fontSize: '20px', fontWeight: '500', mb: '12px'}}>WELCOME</Typography>
-            <Image width={121} height={101} src={logo} placeholder="blur" layout='fixed' />
+            <Typography sx={{fontSize: '20px', fontWeight: '500', mb: '12px', color: '#FFCE00'}}>WELCOME</Typography>
+            <Image width={121} height={91} src={logo} placeholder="blur" layout='fixed' />
             </div>
 
             <div style={{position: 'fixed', left: '0', right: '0', bottom: '0', display: 'grid', margin: '40px'}}>
